refactor: add new package(navigation) and seperate routes
